Padding Tensor Function not passing test due 0th dimension error, needs fixing

+++ mzLib/MachineLearning/AARTNHelperFunctions.cs
@@ -117,6 +117,6 @@ namespace MachineLearning
 
     public class FeaturizedTokens : Tokens 
     {
-        public float[] Features { get; set; }
+        public long[] Features { get; set; }
     }
 }
